Initial interactive plots
The initial interactive plots. Showing the bar plot with 95% confidence interval as error bars.
There is a horizontal line whose value is defined by the user. 
The color of the bar will be changed showing the proportion of data above the horizontal line when changing the line value by clicking